% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{createTSSObject}
\alias{createTSSObject}
\title{Setup gene TSS object}
\usage{
createTSSObject(ensembl_ids, biomart_ensembl)
}
\description{
This creates a dataframe with canonical TSS for each gene that is included in the analysis
}
\examples{
# createTSSObject(c("ENSR00129", "ENSR00139"), biomaRt::useEnsembl(biomart = "genes", dataset = "hsapiens_gene_ensembl", version = 105))
}
\keyword{internal}
