% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts}
\alias{gene_counts}
\title{Example Gene Counts}
\format{
A matrix with 3 rows and 128 columns
\describe{
\item{Rows}{Genes}
\item{Columns}{Samples}
}
}
\source{
University of Michigan
}
\usage{
gene_counts
}
\description{
A matrix containing example gene count data. This is pulled from a larger dataset and was VST transformed.
}
\keyword{datasets}
