% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peak_counts}
\alias{peak_counts}
\title{Example Peak Counts}
\format{
A list of matrices 
\describe{
   \item{H3K27ac}{A matrix of H3K27ac count data with 124 rows and 59 columns} 
   \item{H3K4me1}{A matrix of H3K4me1 count data with 302 rows and 59 columns}
   \item{H3K4me3}{A matrix of H3K4me3 count data with 5 rows and 59 columns}
   \item{ATACSeq}{A matrix of ATACSeq count data with 296 rows and 123 columns}
}
}
\source{
University of Michigan
}
\usage{
peak_counts
}
\description{
A list of matrices containing VST transformed counts for multiple modalities to be integrated. The matrices include features (peaks) as rows and samples as columns. Each matrix contains a "sample_info" attribute that is a dataframe which provides metadata information to be used for data normalization. The rownames should be identical to the column names for the respective matrix. The dataframe should only include numeric and factor data.
}
\keyword{datasets}
