% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAdjustment.R
\name{removeBatchEffect_wrapper}
\alias{removeBatchEffect_wrapper}
\title{Wrapper: Remove Batch Effect
This supplies the experimental model to remove technical variables, while protecting for the experimental variables.}
\usage{
removeBatchEffect_wrapper(
  counts,
  df,
  var_adj,
  var_protect = c("line", "sex", "grp", "train.sed")
)
}
\description{
Wrapper: Remove Batch Effect
This supplies the experimental model to remove technical variables, while protecting for the experimental variables.
}
