% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAdjustment.R
\name{adjustCovariateMatrix}
\alias{adjustCovariateMatrix}
\title{Use of matrix functions to adjust data via linear model}
\usage{
adjustCovariateMatrix(
  counts,
  covariate_df,
  vars_to_protect,
  return_coefficients = FALSE
)
}
\arguments{
\item{counts}{matrix of transformed counts (vst, log, inverse rank normalized, etc.). Should have sample IDs as rownames, feature ID as colnames}

\item{covariate_df}{dataframe with all covariates, sample names should be rownames. should be in the same order as the input matrix}

\item{vars_to_protect}{character vector of column names that should be protected. the coefficients calculated for these variables will not be subtracted from the returned values}

\item{return_coefficients}{logcal. default=FALSE}
}
\value{
matrix of adjusted counts
}
\description{
Use of matrix functions to adjust data via linear model
}
