% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{assignRegion2GeneDirection}
\alias{assignRegion2GeneDirection}
\title{The results from all correlations are used to help assign direction using correlations that have a significant nominal pvalue}
\usage{
assignRegion2GeneDirection(sig_cor_res, all_cor_res, count_mat)
}
\arguments{
\item{sig_cor_res}{dataframe with the significant results to annotate link direction}

\item{all_cor_res}{all resulting correlations}

\item{count_mat}{matrix of counts supplied to correlation analysis}
}
\value{
dataframe with peak-gene link annotation directions with the degree of confidence
}
\description{
Directional assignments are categorized into 3 levels:
\enumerate{
\item Most confident: assign direction using FDR significant links and they are all in agreement
\item Directional disagreement existed, but decision was made by taking the most significant FDR
\item Directional disagreement existed, but decision was made by taking most significant nominal pvalue that was included in comparison
\item These are looser directions associated with the gene - neither distal or promoter region is nominally correlated with the gene, so the direction is based on how that peak is associated with distal peaks that are significantly linked to the gene with high confidence. The iteration of identifying a significant correlation between the peak and another distal peak is divided by 1000 and then added to 4. This shows that the peak:gene relationship exists at confidence level #4, and notes how many distal genes it had to try to correlate with before finding a significant (pval < 0.05) connection.
}
}
