% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{calculateDirectedDistance}
\alias{calculateDirectedDistance}
\title{Genomic distance with the direction}
\usage{
calculateDirectedDistance(subject, target)
}
\value{
A numeric vector of signed distances
}
\description{
This function assigns a negative value if the subject is upstream from the target and positive value if the subject is downstream from the target, while taking into account if the gene is located on the negative or positive strand. This function is created because the GenomicRanges::distance function itself will not account for direction from TSS (upstream (-) or downstream (+)).
}
\examples{
# example code
}
\keyword{internal}
