% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{categorizeAndAnnotateCorrelationResults}
\alias{categorizeAndAnnotateCorrelationResults}
\title{Annotation of genes that are significant and provide some additional downstream information for interpreting linked networks}
\usage{
categorizeAndAnnotateCorrelationResults(.sig_cor_genes, paired_df)
}
\arguments{
\item{.sig_cor_genes}{dataframe that includes genes with significant correlations}

\item{paired_df}{dataframe}
}
\value{
an annotated dataframe
}
\description{
Annotation of genes that are significant and provide some additional downstream information for interpreting linked networks
}
