% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{correlateByChromosome}
\alias{correlateByChromosome}
\title{A wrapper function to carry out correlations in smaller chunks to help save on memory}
\usage{
correlateByChromosome(
  count_mat,
  correlation_pairs,
  grp_contrast = NULL,
  rds_fn = NULL
)
}
\arguments{
\item{count_mat}{numeric matrix with features as columns and samples as rows}

\item{correlation_pairs}{dataframe supplying all pairs of features}

\item{grp_contrast}{a way to filter to include samples from only certain groups}

\item{rds_fn}{filename to save the output to for future access}
}
\value{
a dataframe of full results
}
\description{
A wrapper function to carry out correlations in smaller chunks to help save on memory
}
