% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{corWithDistalPeak}
\alias{corWithDistalPeak}
\title{Iteration of correlations between distal peaks and other peaks residing near it with significant relationships with said gene}
\usage{
corWithDistalPeak(df, full_cor_df, count_mat, iter = 1, iter_batch_cutoff = 5)
}
\arguments{
\item{df}{dataframe of significant correlation results}

\item{full_cor_df}{dataframe of full correlation results}

\item{count_mat}{matrix of counts supplied to correlation analysis}
}
\value{
dataframe with correlation results for that particular iteration
}
\description{
Iteration of correlations between distal peaks and other peaks residing near it with significant relationships with said gene
}
