% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{corWithDistalPeak_wrapper}
\alias{corWithDistalPeak_wrapper}
\title{Wrapper function for corWithDistalPeak to iterate through the correlations}
\usage{
corWithDistalPeak_wrapper(
  full_cor_df,
  count_mat,
  v = TRUE,
  .iter_batch_cutoff = 5
)
}
\arguments{
\item{full_cor_df}{dataframe of full correlation results}

\item{count_mat}{matrix of counts supplied to correlation analysis}

\item{v}{logical. print informative messages}
}
\value{
dataframe with iterated correlation results until a significant correlation was identified with other distal peaks
}
\description{
Wrapper function for corWithDistalPeak to iterate through the correlations
}
