% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{createPeak2GeneObjects}
\alias{createPeak2GeneObjects}
\title{Create Peak-Gene Links
Set up genomicranges objects for the data included in the correlation analysis. Data matrices must have \emph{unique} feature IDs.}
\usage{
createPeak2GeneObjects(gene_counts, peak_counts, biomart_ensembl, ucsc_genome)
}
\description{
Create Peak-Gene Links
Set up genomicranges objects for the data included in the correlation analysis. Data matrices must have \emph{unique} feature IDs.
}
\examples{
# createPeak2GeneObjects(gene_counts, peak_counts, biomaRt::useEnsembl(biomart = "genes", dataset = "rnorvegicus_gene_ensembl", version = 109), "rn7")

}
