% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetup.R
\name{createTSSGr}
\alias{createTSSGr}
\title{Setup gene TSS object}
\usage{
createTSSGr(ensembl_ids, biomart_ensembl, ucsc_genome)
}
\arguments{
\item{ucsc_genome}{genome release formatting in ucsc style. For example: hg38, rn7, etc.}
}
\description{
This creates a dataframe with canonical TSS for each gene that is included in the analysis
}
\examples{
# createTSSObject(c("ENSR00129", "ENSR00139"), biomaRt::useEnsembl(biomart = "genes", dataset = "hsapiens_gene_ensembl", version = 105), "hg38")
}
\keyword{internal}
