% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{editCorResFields}
\alias{editCorResFields}
\title{Annotation of significant promoter peak information}
\usage{
editCorResFields(df)
}
\arguments{
\item{df}{the correlation results}
}
\value{
an annotated dataframe
}
\description{
Annotation of significant promoter peak information
}
