% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{formatMatrixForCorrelation}
\alias{formatMatrixForCorrelation}
\title{Format input}
\usage{
formatMatrixForCorrelation(gene_counts, peak_counts)
}
\arguments{
\item{gene_counts}{count matrix where rownames are sample IDs and colnames are feature ID}
}
\description{
This function pastes the modality at the end of each feature ID and then combines the matrices into one matrix object
}
