% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{matrixCorrelation}
\alias{matrixCorrelation}
\title{Correlation of two matrices for more efficient calculations}
\usage{
matrixCorrelation(
  mat1,
  mat2,
  type = c("pearson", "spearman"),
  rank_data = FALSE,
  return_pvalue = TRUE
)
}
\arguments{
\item{mat1}{first matrix of data to correlate}

\item{mat2}{must have same dimensions as mat1}

\item{type}{method of correlation}

\item{rank_data}{logical}

\item{return_pvalue}{default=TRUE}
}
\value{
A dataframe with five columns containing the pairs of data that were correlated, the correlation coefficient, p-value, and number of samples included in the correlation.
}
\description{
Correlation of two matrices for more efficient calculations
}
\examples{
# example code
}
