% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{processCorrelations}
\alias{processCorrelations}
\title{Process correlation results}
\usage{
processCorrelations(count_mat, cor_res, correlation_pairs, p2g_info)
}
\description{
Genes with a significant correlation with either a promoter or distal peak are retained and focused on further.
}
